+++ src/server/game/Handlers/QuestHandler.cpp
@@ -433,7 +433,7 @@ void WorldSession::HandleQuestLogRemoveQuest(WorldPacket& recvData)
                 // prepare Quest Tracker datas
                 PreparedStatement* stmt = CharacterDatabase.GetPreparedStatement(CHAR_UPD_QUEST_TRACK_ABANDON_TIME);
                 stmt->setUInt32(0, questId);
-                stmt->setUInt32(1, _player->GetGUID().GetCounter());
+                stmt->setUInt64(1, _player->GetGUID().GetCounter());
 
                 // add to Quest Tracker
                 CharacterDatabase.Execute(stmt);
